+++ app/src/main/java/com/example/senior/data/LedRequestBody.kt
@@ -1,6 +1,6 @@
 package com.example.senior.data
 
 data class LedRequestBody(
-    val ledArray: List<Long>,
+    val ledArray: List<Int>,
     val index: Int
 )
